Change Vec2 to template class, changed double precision to float precision

+++ BreakoutApp/src/GameManager.h
@@ -9,5 +9,5 @@ public:
 	bool loadObjects(GameObjects& gObjects);
 
 	// updates all game objects
-	void update(GameObjects& gObjects, GameState& gameState, PaddleMove& paddleMove, double timeStep);
+	void update(GameObjects& gObjects, GameState& gameState, PaddleMove& paddleMove, float timeStep);
 };
